Added hocs (Higher order components) for HeaderBar, refactored directories for screens and added a Switch in the Router
