+++ client/src/App.js
@@ -1,25 +1,19 @@
-import React, { Component, Fragment } from "react";
-import { BrowserRouter as Router, Route } from 'react-router-dom';
-import HeaderBar from "./components/HeaderBar/HeaderBar";
-import CourseRegistration from "./CourseRegistration/CourseRegistration";
+import React, { Component } from "react";
+import { BrowserRouter as Router, Route, Switch } from 'react-router-dom';
+import CourseRegistration from "./screens/CourseRegistration";
+import Login from "./screens/Login"
 import "./App.css";
 
 class App extends Component {
   render() {
     return (
-      <div className="app">
-        <HeaderBar />
-        <div className="app-content">
-          <Router>
-            <Fragment>
-              <Route exact path="/" render={ () => 
-                <p>Hello World!</p>
-              }/>
-              <Route path="/courseregistration" component={ CourseRegistration }/>            
-            </Fragment>
-          </Router>          
-        </div>
-      </div>
+      <Router>
+        <Switch>
+          <Route exact path="/login" component={Login} />
+          <Route path="/courseregistration" component={CourseRegistration} />
+        </Switch>
+      </Router>
+
     );
   }
 }
