+++ client/src/components/HeaderBar/index.js
@@ -1,0 +1,1 @@
+export { default } from "./HeaderBar.jsx"
