+++ client/src/screens/Login/Login.js
@@ -1,0 +1,25 @@
+import React, { Component, Fragment } from "react";
+import "./Login.css"
+
+class Login extends Component {
+	render() {
+		return (
+			<div className="login">
+				<Fragment>
+					<div className="header">
+						<div className="title"> Project Athena</div>
+					</div>
+				</Fragment>
+
+				<div className="content">
+					<div className="description">
+					</div>
+					<div className="form">
+					</div>
+				</div>
+			</div>
+		);
+	}
+}
+
+export default Login;
